back using session

+++ commerce/pages/api/admin/users/[id].ts
@@ -1,14 +1,14 @@
-// import { getSession } from "next-auth/react";
+import { getSession } from "next-auth/react";
 import User from "@/models/User";
 import db from "@/utils/db";
 
 
 const handler = async (req:any, res:any) => {
-    // const session:any = await getSession({req});
-    // if (!session || !session.user.isAdmin) {
-    //      return res.status(401).send('admin signin required');
+    const session:any = await getSession({req});
+    if (!session || !session.user.isAdmin) {
+         return res.status(401).send('admin signin required');
 
-    // }
+    }
     if(req.method === 'DELETE'){
         return deleteHandler(req, res);
